+++ performance/src/main/java/com/ticketing/performance/application/service/PerformanceService.java
@@ -13,10 +13,11 @@ import com.ticketing.performance.common.util.SecurityUtil;
 import com.ticketing.performance.domain.model.Performance;
 import com.ticketing.performance.domain.model.SeatStatus;
 import com.ticketing.performance.domain.repository.PerformanceRepository;
-import com.ticketing.performance.presentation.dto.PerformanceSearchRequestDto;
+import com.ticketing.performance.presentation.dto.performance.PerformanceSearchRequestDto;
 import com.ticketing.performance.presentation.dto.performance.CreatePrfRequestDto;
 import com.ticketing.performance.presentation.dto.performance.UpdatePrfRequestDto;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
