+++ performance/src/main/java/com/ticketing/performance/presentation/controller/PerformanceController.java
@@ -7,7 +7,7 @@ import com.ticketing.performance.application.dto.performance.PrfListResponseDto;
 import com.ticketing.performance.application.dto.performance.UpdatePrfResponseDto;
 import com.ticketing.performance.application.service.PerformanceService;
 import com.ticketing.performance.common.response.CommonResponse;
-import com.ticketing.performance.presentation.dto.PerformanceSearchRequestDto;
+import com.ticketing.performance.presentation.dto.performance.PerformanceSearchRequestDto;
 import com.ticketing.performance.presentation.dto.performance.CreatePrfRequestDto;
 import com.ticketing.performance.presentation.dto.performance.UpdatePrfRequestDto;
 import lombok.RequiredArgsConstructor;
